Type <-> Column description mapping improvements
* Better automatic column precision when scale is set but length is not specified
* Filter fraction scale option independantly of length
* Manually set PostgreSQL "nummeric" type max length to 1000. See [documentation](https://www.postgresql.org/docs/13/datatype-numeric.html#DATATYPE-NUMERIC-DECIMAL)

ns-xml SQL schema 2.1 support

* Handle new sql:integer data type

TYPE_DEFAULT_SCALE type property

Add new type property to indicates that a type will have a implicit fraction scale when a column is declared without precision and fraction scale.

Example in MySQL
```
f DECIMAL
```
`f` will have a fraction scale of 0

* Handle TYPE_DEFAULT_SCALE in TypeRegistry and MySQL
* Update tests to check float/fixed precision numeric types

+++ src/Structure/Traits/XMLStructureFileTrait.php
@@ -189,7 +189,7 @@ trait XMLStructureFileTrait
 			K::DATATYPE_TIME => 'timestamp',
 			K::DATATYPE_NUMBER => 'numeric',
 			K::DATATYPE_FLOAT => 'numeric',
-			K::DATATYPE_INTEGER => 'numeric',
+			K::DATATYPE_INTEGER => 'integer',
 			K::DATATYPE_NULL => 'null',
 			K::DATATYPE_STRING => 'string',
 			K::DATATYPE_UNDEFINED => 'string'
